+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 In Development
 --------------
 
+* A report of box activity, including sent/received mail totals and logins by user, is now emailed to the box's administrator user each week.
 * Update Roundcube to version 1.3.4 and Z-Push to version 2.3.9.
 * The undocumented feature for proxying web requests to another server now sets X-Forwarded-For.